Updates to Web App
Dropdown menu was changed in the index.html file to a tick box selection, so that multiple azure containers can be selected and used in a search. This will simulate choosing multiple base stations.

routes.py file was modified to take in multiple containers when updating a map.

start_search function was added for users to begin a search and for data within that search be saved
end_search function added to end a search and collect search data to be packaged into GPX format.
download_gpx function creates a download link for users to click on.

All subject to improvement and not final.