Improved function to Upload data to database and to retrieve data from database
Created and modified SQL database schema to allow for JSON data to be stored in the gps_JSON column.
gps_data from each base station is decoded and converted into a string for legal storage in database.

When a search ends, the data is uploaded to the SQL database through the upload_search_data(). For each base station that was involved in the search, it INSERTS a new row into the database table.

I have also added a UI feature that allows users to select a historical search by selecting the start date, end date and base stations involved in the search. This is processed in the get_historical_searches function, which runs a query that gets all rows that fit the filter criteria. For now this is returned as a list of tuples, each tuple representing a row in the database.

Also, the UI feature to select base stations has been set so that users can only select base stations that are currently stored in the database. This renders in when the web app page is rendered.